requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
jinja2==3.1.4
